refactor: extract shared ErrorCard component
Merge the near-identical floating-card shell from error.tsx, global-error.tsx,
and not-found.tsx into src/components/shared/ErrorCard.tsx. Each page keeps
its own icon, copy, and action buttons via props; global-error.tsx keeps its
<html><body><MotionConfig> wrapper outside the shared component since that's
a root-layout concern.

Partially addresses #1474
